[OCMS - Level 3] Updated is_late.

+++ Back-End/OCMS - Level 1,2/plugins/app/latearrival/updates/add_late_column.php
@@ -15,8 +15,10 @@ class AddLateColumn extends Migration
 
     public function down()
     {
-        Schema::table('app_arrival_arrivals', function (Blueprint $table) {
-            $table->dropColumn('is_late');
-        });
+        if (Schema::hasColumn('app_arrival_arrivals', 'is_late')) {
+            Schema::table('app_arrival_arrivals', function (Blueprint $table) {
+                $table->dropColumn('is_late');
+            });
+        }
     }
 }
